src/userservice/requirements.txt
bcrypt==4.0.1
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.12
psycopg2-binary==2.9.5
six==1.16.0
typing_extensions==4.4.0
pytest==8.3.2
